Corrigido problemas com GraphQuerier e documentação
