+++ module/README.md
@@ -1,0 +1,6 @@
+To run the Acess Point, you will need to install and start Neo4j Desktop from its [website](https://neo4j.com/download-center/#desktop), and after that:
+
+- Create a new project
+- Add a "Local DBMS"
+  - Use for password: "password"
+- Start the server
